+++ tests/UnitTests/MqttEngineTests.cs
@@ -3,196 +3,22 @@ using MQTTnet;
 using MQTTnet.Protocol;
 using Xunit;
 using CrowsNestMqtt.BusinessLogic;
-using CrowsNestMqtt.BusinessLogic.Configuration; // Added for TopicBufferLimit
+using CrowsNestMqtt.BusinessLogic.Configuration;
 using System.Linq;
 
 namespace CrowsNestMqtt.UnitTests;
 
-public class MqttEngineTests : IClassFixture<MqttBrokerFixture>
+/// <summary>
+/// Pure unit tests for MqttEngine that do NOT require a running MQTT broker.
+/// Broker-dependent tests have been moved to Integration.Tests/MqttEngineIntegrationTests.cs.
+/// </summary>
+public class MqttEngineTests
 {
-    private readonly MqttBrokerFixture _brokerFixture;
-    private readonly ITestOutputHelper _output;
-    private static readonly TimeSpan TestConnectTimeout = TimeSpan.FromSeconds(30);
-
-    public MqttEngineTests(MqttBrokerFixture brokerFixture, ITestOutputHelper output)
-    {
-        _brokerFixture = brokerFixture;
-        _output = output;
-    }
-
-    private void LogBrokerState(string testName)
-    {
-        _output.WriteLine($"[{testName}] Broker fixture: IsRunning={_brokerFixture.IsRunning}, Port={_brokerFixture.Port}, Host={_brokerFixture.Hostname}");
-        if (_brokerFixture.StartupError != null)
-            _output.WriteLine($"[{testName}] Broker startup error: {_brokerFixture.StartupError}");
-    }
-
-    private void AttachEngineLogging(MqttEngine engine, string testName)
-    {
-        engine.LogMessage += (_, msg) => _output.WriteLine($"[{testName}] Engine: {msg}");
-        engine.ConnectionStateChanged += (_, args) => _output.WriteLine($"[{testName}] ConnectionState: IsConnected={args.IsConnected}, Status={args.ConnectionStatus}, Error={args.ErrorMessage}");
-    }
-    
     [Fact]
     public void MessageReceived_BatchEvent_Should_Fire_Once_Per_Batch()
     {
         // This test is a placeholder for the new batch event behavior.
-        // It should be updated after the refactor to verify that the batch event
-        // is fired only once per batch, even with a high volume of messages.
-        // The implementation will depend on the new event signature.
         Assert.True(true, "Batch event test placeholder. Update after refactor.");
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task MqttEngine_Should_Receive_Published_Message()
-    {
-        const string testName = nameof(MqttEngine_Should_Receive_Published_Message);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        string brokerHost = _brokerFixture.Hostname;
-        int brokerPort = _brokerFixture.Port;
-        var connectionSettings = new MqttConnectionSettings
-        {
-            Hostname = brokerHost,
-            Port = brokerPort,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(connectionSettings);
-        AttachEngineLogging(engine, testName);
-
-        IdentifiedMqttApplicationMessageReceivedEventArgs? receivedArgs = null;
-        using var messageReceivedEvent = new ManualResetEventSlim(false);
-
-engine.MessagesBatchReceived += (sender, batch) =>
-        {
-            foreach (var args in batch)
-            {
-                if (args.ApplicationMessage.Topic == "test/topic")
-                {
-                    receivedArgs = args;
-                    messageReceivedEvent.Set();
-                    break;
-                }
-            }
-        };
-
-        await engine.ConnectAsync(cts.Token);
-
-        // Act: Publish a test message using a publisher client.
-        var factory = new MqttClientFactory();
-        var publisher = factory.CreateMqttClient();
-        var publisherOptions = new MqttClientOptionsBuilder()
-            .WithTcpServer(brokerHost, brokerPort)
-            .WithCleanSession(true)
-            .Build();
-
-        await publisher.ConnectAsync(publisherOptions, cts.Token);
-
-        var payload = "Test Message";
-        var message = new MqttApplicationMessageBuilder()
-            .WithTopic("test/topic")
-            .WithPayload(payload)
-            .WithQualityOfServiceLevel(MqttQualityOfServiceLevel.AtLeastOnce)
-            .Build();
-
-        await publisher.PublishAsync(message, cts.Token);
-
-        // Wait for the message to be received
-        if (!messageReceivedEvent.Wait(TimeSpan.FromSeconds(10), cts.Token))
-        {
-            Assert.Fail("Timeout waiting for message.");
-        }
-
-        // Assert
-        Assert.NotNull(receivedArgs);
-        Assert.Equal("test/topic", receivedArgs.ApplicationMessage.Topic);
-        Assert.Equal(payload, receivedArgs.ApplicationMessage.ConvertPayloadToString());
-
-        // Cleanup
-        await publisher.DisconnectAsync(new MqttClientDisconnectOptions(), CancellationToken.None);
-        await engine.DisconnectAsync(CancellationToken.None);
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task MqttEngine_Should_Handle_Empty_Payload_Message()
-    {
-        const string testName = nameof(MqttEngine_Should_Handle_Empty_Payload_Message);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        string brokerHost = _brokerFixture.Hostname;
-        int brokerPort = _brokerFixture.Port;
-        var connectionSettings = new MqttConnectionSettings
-        {
-            Hostname = brokerHost,
-            Port = brokerPort,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(connectionSettings);
-        AttachEngineLogging(engine, testName);
-
-        IdentifiedMqttApplicationMessageReceivedEventArgs? receivedArgs = null;
-        using var messageReceivedEvent = new ManualResetEventSlim(false);
-
-engine.MessagesBatchReceived += (sender, batch) =>
-        {
-            foreach (var args in batch)
-            {
-                if (args.ApplicationMessage.Topic == "test/empty_payload_topic")
-                {
-                    receivedArgs = args;
-                    messageReceivedEvent.Set();
-                    break;
-                }
-            }
-        };
-
-        await engine.ConnectAsync(cts.Token);
-
-        // Act: Publish a test message with an empty payload.
-        var factory = new MqttClientFactory();
-        var publisher = factory.CreateMqttClient();
-        var publisherOptions = new MqttClientOptionsBuilder()
-            .WithTcpServer(brokerHost, brokerPort)
-            .WithCleanSession(true)
-            .Build();
-
-        await publisher.ConnectAsync(publisherOptions, cts.Token);
-
-        var message = new MqttApplicationMessageBuilder()
-            .WithTopic("test/empty_payload_topic")
-            .WithPayload(Array.Empty<byte>()) // Empty payload
-            .WithQualityOfServiceLevel(MqttQualityOfServiceLevel.AtLeastOnce)
-            .Build();
-
-        await publisher.PublishAsync(message, cts.Token);
-
-        // Wait for the message to be received
-        if (!messageReceivedEvent.Wait(TimeSpan.FromSeconds(10), cts.Token))
-        {
-            Assert.Fail("Timeout waiting for message with empty payload.");
-        }
-
-        // Assert
-        Assert.NotNull(receivedArgs);
-        Assert.Equal("test/empty_payload_topic", receivedArgs.ApplicationMessage.Topic);
-        // Payload is ReadOnlySequence<byte>, a struct, so it cannot be null.
-        // We check IsEmpty or Length instead.
-        Assert.True(receivedArgs.ApplicationMessage.Payload.IsEmpty, "Payload should be empty.");
-        Assert.Null(receivedArgs.ApplicationMessage.ConvertPayloadToString());
-
-        // Cleanup
-        await publisher.DisconnectAsync(new MqttClientDisconnectOptions(), CancellationToken.None);
-        await engine.DisconnectAsync(CancellationToken.None);
     }
 
     [Fact]
@@ -308,44 +134,6 @@ engine.MessagesBatchReceived += (sender, batch) =>
     }
 
     [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task SubscribeAsync_WhenConnected_ShouldSucceed()
-    {
-        const string testName = nameof(SubscribeAsync_WhenConnected_ShouldSucceed);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(settings);
-        AttachEngineLogging(engine, testName);
-
-        try
-        {
-            await engine.ConnectAsync(cts.Token);
-
-            // Act
-            var result = await engine.SubscribeAsync("test/subscribe/topic", MqttQualityOfServiceLevel.AtLeastOnce);
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.Single(result.Items);
-            Assert.Equal("test/subscribe/topic", result.Items.First().TopicFilter.Topic);
-        }
-        finally
-        {
-            await engine.DisconnectAsync(CancellationToken.None);
-        }
-    }
-
-    [Fact]
     public async Task UnsubscribeAsync_WhenNotConnected_ShouldThrowInvalidOperationException()
     {
         // Arrange
@@ -361,112 +149,6 @@ engine.MessagesBatchReceived += (sender, batch) =>
             () => engine.UnsubscribeAsync("test/topic"));
         
         Assert.Equal("Client is not connected.", exception.Message);
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task UnsubscribeAsync_WhenConnected_ShouldSucceed()
-    {
-        const string testName = nameof(UnsubscribeAsync_WhenConnected_ShouldSucceed);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(settings);
-        AttachEngineLogging(engine, testName);
-
-        try
-        {
-            await engine.ConnectAsync(cts.Token);
-            
-            // First subscribe to a topic
-            await engine.SubscribeAsync("test/unsubscribe/topic");
-
-            // Act
-            var result = await engine.UnsubscribeAsync("test/unsubscribe/topic");
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.Single(result.Items);
-            Assert.Equal("test/unsubscribe/topic", result.Items.First().TopicFilter);
-        }
-        finally
-        {
-            await engine.DisconnectAsync(CancellationToken.None);
-        }
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task PublishAsync_WhenNotConnected_ShouldLogWarningAndReturn()
-    {
-        const string testName = nameof(PublishAsync_WhenNotConnected_ShouldLogWarningAndReturn);
-        LogBrokerState(testName);
-
-        // Arrange
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port
-        };
-        using var engine = new MqttEngine(settings);
-        
-        string? logMessage = null;
-        engine.LogMessage += (sender, message) => logMessage = message;
-
-        // Act
-        await engine.PublishAsync("test/topic", "test payload");
-
-        // Assert
-        Assert.Equal("Cannot publish: Client is not connected.", logMessage);
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task PublishAsync_WhenConnected_ShouldSucceed()
-    {
-        const string testName = nameof(PublishAsync_WhenConnected_ShouldSucceed);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(settings);
-        AttachEngineLogging(engine, testName);
-        
-        string? logMessage = null;
-        engine.LogMessage += (sender, message) => logMessage = message;
-
-        try
-        {
-            await engine.ConnectAsync(cts.Token);
-
-            // Act
-            await engine.PublishAsync("test/publish/topic", "test payload", 
-                retain: true, qos: MqttQualityOfServiceLevel.AtLeastOnce);
-
-            // Assert
-            Assert.Contains("Successfully published to 'test/publish/topic'", logMessage);
-        }
-        finally
-        {
-            await engine.DisconnectAsync(CancellationToken.None);
-        }
     }
 
     [Fact]
@@ -488,82 +170,6 @@ engine.MessagesBatchReceived += (sender, batch) =>
     }
 
     [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task GetMessagesForTopic_AfterReceivingMessage_ShouldReturnMessages()
-    {
-        const string testName = nameof(GetMessagesForTopic_AfterReceivingMessage_ShouldReturnMessages);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(settings);
-        AttachEngineLogging(engine, testName);
-
-        using var messageReceived = new ManualResetEventSlim(false);
-engine.MessagesBatchReceived += (sender, batch) =>
-        {
-            foreach (var args in batch)
-            {
-                if (args.ApplicationMessage.Topic == "test/getmessages/topic")
-                {
-                    messageReceived.Set();
-                    break;
-                }
-            }
-        };
-
-        try
-        {
-            await engine.ConnectAsync(cts.Token);
-
-            // Publish a message to create buffer content
-            var factory = new MqttClientFactory();
-            var publisher = factory.CreateMqttClient();
-            var publisherOptions = new MqttClientOptionsBuilder()
-                .WithTcpServer(_brokerFixture.Hostname, _brokerFixture.Port)
-                .WithCleanSession(true)
-                .Build();
-
-            await publisher.ConnectAsync(publisherOptions, cts.Token);
-
-            var message = new MqttApplicationMessageBuilder()
-                .WithTopic("test/getmessages/topic")
-                .WithPayload("test message content")
-                .WithQualityOfServiceLevel(MqttQualityOfServiceLevel.AtLeastOnce)
-                .Build();
-
-            await publisher.PublishAsync(message, cts.Token);
-
-            // Wait for message to be received and buffered
-            Assert.True(messageReceived.Wait(TimeSpan.FromSeconds(10)), "Message was not received within timeout");
-
-            // Act
-            var bufferedMessages = engine.GetMessagesForTopic("test/getmessages/topic");
-
-            // Assert
-            Assert.NotNull(bufferedMessages);
-            var messagesList = bufferedMessages.ToList();
-            Assert.Single(messagesList);
-Assert.Equal("test/getmessages/topic", messagesList[0].Message.Topic);
-Assert.Equal("test message content", messagesList[0].Message.ConvertPayloadToString());
-
-            await publisher.DisconnectAsync(new MqttClientDisconnectOptions(), CancellationToken.None);
-        }
-        finally
-        {
-            await engine.DisconnectAsync(CancellationToken.None);
-        }
-    }
-
-    [Fact]
     public void GetBufferedTopics_WhenNoMessages_ShouldReturnEmptyCollection()
     {
         // Arrange
@@ -580,99 +186,6 @@ Assert.Equal("test message content", messagesList[0].Message.ConvertPayloadToStr
         // Assert
         Assert.NotNull(topics);
         Assert.Empty(topics);
-    }
-
-    [Fact]
-    [Trait("Category", "RequiresMqttBroker")]
-    public async Task GetBufferedTopics_AfterReceivingMessages_ShouldReturnTopicList()
-    {
-        const string testName = nameof(GetBufferedTopics_AfterReceivingMessages_ShouldReturnTopicList);
-        LogBrokerState(testName);
-        Assert.True(_brokerFixture.IsRunning, "Embedded MQTT broker is not running.");
-
-        // Arrange
-        using var cts = new CancellationTokenSource(TestConnectTimeout);
-        var settings = new MqttConnectionSettings
-        {
-            Hostname = _brokerFixture.Hostname,
-            Port = _brokerFixture.Port,
-            ClientId = $"test-client-{Guid.NewGuid()}",
-            CleanSession = true
-        };
-        using var engine = new MqttEngine(settings);
-        AttachEngineLogging(engine, testName);
-
-        var messagesReceived = 0;
-        using var messageReceivedEvent = new ManualResetEventSlim(false);
-engine.MessagesBatchReceived += (sender, batch) =>
-        {
-            foreach (var args in batch)
-            {
-                // Only count messages for our test topics
-                if (args.ApplicationMessage.Topic == "test/buffered/topic1" ||
-                    args.ApplicationMessage.Topic == "test/buffered/topic2")
-                {
-                    Interlocked.Increment(ref messagesReceived);
-                    if (messagesReceived >= 2) // Wait for both test messages
-                    {
-                        messageReceivedEvent.Set();
-                    }
-                }
-            }
-        };
-
-        try
-        {
-            await engine.ConnectAsync(cts.Token);
-
-            // The engine automatically subscribes to "#" so no explicit subscription needed
-            // Give a small delay to ensure automatic subscription is active
-            await Task.Delay(1000);
-
-            // Publish messages to different topics
-            var factory = new MqttClientFactory();
-            var publisher = factory.CreateMqttClient();
-            var publisherOptions = new MqttClientOptionsBuilder()
-                .WithTcpServer(_brokerFixture.Hostname, _brokerFixture.Port)
-                .WithCleanSession(true)
-                .Build();
-
-            await publisher.ConnectAsync(publisherOptions, cts.Token);
-
-            var message1 = new MqttApplicationMessageBuilder()
-                .WithTopic("test/buffered/topic1")
-                .WithPayload("message 1")
-                .WithQualityOfServiceLevel(MqttQualityOfServiceLevel.AtLeastOnce)
-                .Build();
-
-            var message2 = new MqttApplicationMessageBuilder()
-                .WithTopic("test/buffered/topic2")
-                .WithPayload("message 2")
-                .WithQualityOfServiceLevel(MqttQualityOfServiceLevel.AtLeastOnce)
-                .Build();
-
-            await publisher.PublishAsync(message1, cts.Token);
-            await publisher.PublishAsync(message2, cts.Token);
-
-            // Wait for messages to be received and buffered
-            Assert.True(messageReceivedEvent.Wait(TimeSpan.FromSeconds(15)), "Messages were not received within timeout");
-
-            // Act
-            var topics = engine.GetBufferedTopics().ToList();
-
-            // Assert
-            Assert.NotNull(topics);
-            Assert.Contains("test/buffered/topic1", topics);
-            Assert.Contains("test/buffered/topic2", topics);
-            // Note: There might be other topics from previous test runs or retained messages
-            Assert.True(topics.Count >= 2, $"Expected at least 2 topics, but found {topics.Count}");
-
-            await publisher.DisconnectAsync(new MqttClientDisconnectOptions(), CancellationToken.None);
-        }
-        finally
-        {
-            await engine.DisconnectAsync(CancellationToken.None);
-        }
     }
 
     [Fact]
